Place everything under a resources
