+++ lib/ncs_analytics/client.rb
@@ -1,54 +1,54 @@
-require 'ncs_analytics/resource/harvest'
-require 'ncs_analytics/resource/item'
-require 'ncs_analytics/resource/package'
-require 'ncs_analytics/resource/plant_batch'
-require 'ncs_analytics/resource/plant'
-require 'ncs_analytics/resource/room'
-require 'ncs_analytics/resource/sale'
-require 'ncs_analytics/resource/strain'
-require 'ncs_analytics/resource/transfer'
-require 'ncs_analytics/resource/test'
+require 'ncs_analytics/resources/harvest'
+require 'ncs_analytics/resources/item'
+require 'ncs_analytics/resources/package'
+require 'ncs_analytics/resources/plant_batch'
+require 'ncs_analytics/resources/plant'
+require 'ncs_analytics/resources/room'
+require 'ncs_analytics/resources/sale'
+require 'ncs_analytics/resources/strain'
+require 'ncs_analytics/resources/transfer'
+require 'ncs_analytics/resources/test'
 
 module NcsAnalytics
   class Client
     def harvest
-      @harvest ||= NcsAnalytics::Harvest.new
+      @harvest ||= NcsAnalytics::Resources::Harvest.new
     end
 
     def item
-      @item ||= NcsAnalytics::Item.new
+      @item ||= NcsAnalytics::Resources::Item.new
     end
 
     def package
-      @package ||= NcsAnalytics::Package.new
+      @package ||= NcsAnalytics::Resources::Package.new
     end
 
     def plant_batch
-      @plant_batch ||= NcsAnalytics::PlantBatch.new
+      @plant_batch ||= NcsAnalytics::Resources::PlantBatch.new
     end
 
     def plant
-      @plant ||= NcsAnalytics::Plant.new
+      @plant ||= NcsAnalytics::Resources::Plant.new
     end
 
     def room
-      @room ||= NcsAnalytics::Room.new
+      @room ||= NcsAnalytics::Resources::Room.new
     end
 
     def sale
-      @sale ||= NcsAnalytics::Sale.new
+      @sale ||= NcsAnalytics::Resources::Sale.new
     end
 
     def strain
-      @strain ||= NcsAnalytics::Strain.new
+      @strain ||= NcsAnalytics::Resources::Strain.new
     end
 
     def transfer
-      @transfer ||= NcsAnalytics::Transfer.new
+      @transfer ||= NcsAnalytics::Resources::Transfer.new
     end
 
     def test
-      @test ||= NcsAnalytics::Test.new
+      @test ||= NcsAnalytics::Resources::Test.new
     end
   end
 end
