--- lib/ncs_analytics/resource/sale.rb
@@ -1,13 +1,0 @@
-module NcsAnalytics
-  class Sale < NcsAnalytics::Resource
-    def all
-      get(:receipts)
-    end
-
-    def create(payload)
-      # TODO: Validate the payload
-      payload = [payload] unless payload.is_a?(Array)
-      post(:receipts, payload)
-    end
-  end
-end
