+++ spec/ncs_analytics/resources/harvest_spec.rb
@@ -1,6 +1,6 @@
 require 'spec_helper'
 
-describe NcsAnalytics::Harvest do
+describe NcsAnalytics::Resources::Harvest do
   subject { described_class.new }
 
   before { configure_client }
